refactor(web): 团队三个面板组件搬出 Channel.tsx 到 components/team/（PR A，零行为变化）
