+++ web/src/pages/Channel.tsx
@@ -1,11 +1,15 @@
 // 频道页：presence 条 + 实时消息流 + 内联错误条幅 + 插话框。
 // App 用 key={slug} 挂载本组件，切频道即整体重建（socket/状态零残留）。
-import { Fragment, useCallback, useEffect, useLayoutEffect, useMemo, useReducer, useRef, useState } from "react";
+import { useCallback, useEffect, useLayoutEffect, useMemo, useReducer, useRef, useState } from "react";
 import { guessJoinPackHarness } from "../lib/joinPack";
 import type { CSSProperties, ReactNode } from "react";
-import { buildHostBoard, type Attachment, type ChannelSquad, type CollaborationRole, type MsgFrame, type ParticipantRemovedFrame, type PresenceEntry, type PublicDirectedDelivery, type ReadCursor, type SearchHit, type Sender, type TaskAssigneeKind, type TaskRecord, type TaskState, type TaskSummary, type WakeDelivery } from "@agentparty/shared";
+import { buildHostBoard, type Attachment, type ChannelSquad, type MsgFrame, type ParticipantRemovedFrame, type PresenceEntry, type PublicDirectedDelivery, type ReadCursor, type SearchHit, type Sender, type TaskAssigneeKind, type TaskRecord, type TaskState, type TaskSummary, type WakeDelivery } from "@agentparty/shared";
 import { AgentDetailPanel } from "../components/AgentDetailModal";
 import { TeamTabs } from "../components/TeamTabs";
+import { DivisionBoard } from "../components/team/DivisionBoard";
+import { TeamPanel } from "../components/team/TeamPanel";
+import { AgentBoardPanel } from "../components/team/AgentBoardPanel";
+import { teamMemberOnlineNames } from "../components/team/agentBoard";
 import { AgentJoin, type JoinGuideSession } from "../components/AgentJoin";
 import { AgentTokens } from "../components/AgentTokens";
 import { VisibilityToggle } from "../components/VisibilityToggle";
@@ -34,7 +38,6 @@ import {
 } from "../lib/channelNavigation";
 import { useAuthoritativePendingDecisions } from "../lib/pendingDecisions";
 import { AttachmentList } from "../components/AttachmentList";
-import { OrgTreePreview } from "../components/OrgTreePreview";
 import {
   addChannelMember,
   archiveChannel,
@@ -86,9 +89,7 @@ import { mentionCandidates, parseDraftMentions, type DraftMentionStatus } from "
 import { buildReceipts, type MentionReceipt } from "../lib/wakeReceipt";
 import { completionMessages } from "../lib/completions";
 import { catchupKey, shouldShowCatchup, summarizeCatchup } from "../lib/digest";
-import { buildOrgTree, type OrgMemberInput } from "../lib/orgTree";
-import { formatDivisionSection, mergeDivisionIntoCharter, type DivisionCharterRole } from "../lib/divisionCharter";
-import { declaredAgentRoles, formatAgentRoleSummary } from "../lib/divisionSummary";
+import { formatAgentRoleSummary } from "../lib/divisionSummary";
 import {
   clearDeliveredMentionNotifications,
   isDesktopRuntime,
@@ -119,16 +120,16 @@ import { useCharterRequestGeneration } from "../lib/useCharterRequestGeneration"
 import { isNearBottom, pinToBottom } from "../lib/scrollPin";
 import { summarizeReplyPreview } from "../lib/replyPreview";
 import { fmtTime } from "../lib/time";
-import { groupTeamMessages, summarizeTeams, type TeamMessageThread, type TeamSummary } from "../lib/teams";
+import { groupTeamMessages, summarizeTeams, type TeamMessageThread } from "../lib/teams";
 import { resolveTeamMemberView } from "../lib/teamMember";
+import { roleDraftFrom, teamRoleBuckets, type RoleDraft } from "../lib/channelRoles";
 import { ChannelSocket } from "../lib/ws";
 import { channelReducer, initialChannelState } from "../state";
-import { useT, type TFunc } from "../i18n/useT";
+import { useT } from "../i18n/useT";
 import { ChannelToolstrip } from "../components/ChannelToolstrip";
 import "../i18n/strings/Channel";
 import "../i18n/strings/Composer";
 import "../i18n/strings/WakeReceipt";
-import { presenceStateLabel, residencyLabel } from "../lib/presenceLabels";
 
 const EMPTY_RECENT_MESSAGES: MsgFrame[] = [];
 const SCREEN_READER_ONLY_STYLE: CSSProperties = {
@@ -177,7 +178,6 @@ interface Props {
 // IM 式加载：初始/上翻每页条数，与 DOM 消息窗口上限（贴底时超出即丢最老页，上翻可拉回）
 const PAGE_SIZE = 50;
 const MESSAGE_CAP = 300;
-const COLLAB_ROLES: CollaborationRole[] = ["host", "worker", "reviewer", "observer"];
 // 触顶阈值：滚动到离顶部这么近就预取上一页
 const TOP_LOAD_PX = 80;
 
@@ -235,11 +235,6 @@ function writeSeenCharterRev(slug: string, rev: number) {
   }
 }
 
-export interface RoleDraft {
-  role: CollaborationRole;
-  responsibility: string;
-}
-
 type ChannelPanel = "charter" | "team" | "focus" | "tasks" | "search" | "admin";
 type MemberDetailSource = Extract<ChannelPanel, "team" | "admin">;
 
@@ -269,19 +264,6 @@ function compactTaskTitle(text: string, fallback: string): string {
   const raw = text.replace(/\s+/g, " ").trim();
   const label = raw === "" ? fallback : raw;
   return label.length > 120 ? `${label.slice(0, 117)}...` : label;
-}
-
-function roleDraftFrom(role: ChannelRoleInfo): RoleDraft {
-  return { role: role.role, responsibility: role.responsibility ?? "" };
-}
-
-function roleViewFor(role: ChannelRoleInfo, identity: ChannelIdentity | undefined, t: TFunc) {
-  const kind = role.kind ?? identity?.kind ?? "agent";
-  const account = role.account ?? identity?.account;
-  const display = role.display ?? identity?.display ?? (kind === "human" && account ? account : role.name);
-  const accountLabel = account && account !== "" ? account : kind === "human" ? display : t("Channel.roles.unowned");
-  const owner = account && account !== display ? account : null;
-  return { role, display, accountLabel, owner, kind };
 }
 
 export interface GuardSettingsPanelProps {
@@ -417,118 +399,6 @@ export function GuardSettingsPanel({
   );
 }
 
-function roleCountLabel(role: CollaborationRole, count: number, t: TFunc): string {
-  return t("Channel.roles.roleCount", { role, count: String(count) });
-}
-
-function selfReportedRoles(
-  assignedRoles: ChannelRoleInfo[],
-  presence: Record<string, PresenceEntry>,
-  identities: ChannelIdentity[],
-): ChannelRoleInfo[] {
-  const assigned = new Set(assignedRoles.map((role) => role.name));
-  const identityByName = new Map(identities.map((identity) => [identity.name, identity]));
-  const roles: ChannelRoleInfo[] = [];
-  for (const [name, entry] of Object.entries(presence)) {
-    if (assigned.has(name)) continue;
-    if (entry.role_source !== "self") continue;
-    if (entry.role === undefined || !COLLAB_ROLES.includes(entry.role)) continue;
-    const identity = identityByName.get(name);
-    const kind = entry.kind ?? identity?.kind;
-    const account = entry.account ?? identity?.account;
-    roles.push({
-      name,
-      role: entry.role,
-      responsibility: entry.note && entry.note.trim() !== "" ? entry.note : null,
-      assigned_by: name,
-      assigned_at: entry.ts ?? entry.last_seen ?? 0,
-      ...(kind === undefined ? {} : { kind }),
-      ...(account === undefined ? {} : { account }),
-      display: identity?.display ?? name,
-    });
-  }
-  return roles;
-}
-
-interface UnassignedMember {
-  name: string;
-  display: string;
-  accountLabel: string;
-  owner: string | null;
-  kind: Sender["kind"];
-}
-
-// issue #169：分工面板此前只收录「已分配角色」（roles）+「self-report 过角色」
-// （presence role_source==="self"）的成员——已连接但从没声明过角色的 agent 会被
-// 整条跳过，界面上直接消失（"频道四个 agent 分工面板只有两个"）。这里把他们也
-// 收进名单，用「未分工」占位展示，而不是从 roster 里彻底丢失。
-// 名单来源取 participants（当前连接）∪ presence（当前/最近连接过）∪ identities
-// （channel 曾经见过的身份）的并集，与 Team 头部的权威 roster 口径一致；已在
-// roles/selfRoles 里出现的名字（assigned 或 self）不重复收录。
-function unassignedMembers(
-  assignedRoles: ChannelRoleInfo[],
-  selfRoles: ChannelRoleInfo[],
-  presence: Record<string, PresenceEntry>,
-  identities: ChannelIdentity[],
-  participants: Sender[],
-  t: TFunc,
-): UnassignedMember[] {
-  const known = new Set([...assignedRoles, ...selfRoles].map((role) => role.name));
-  const identityByName = new Map(identities.map((identity) => [identity.name, identity]));
-  const participantByName = new Map(participants.map((participant) => [participant.name, participant]));
-  const names = new Set([
-    ...Object.keys(presence),
-    ...identities.map((identity) => identity.name),
-    ...participants.map((participant) => participant.name),
-  ]);
-  const members: UnassignedMember[] = [];
-  for (const name of names) {
-    if (name === "system" || known.has(name)) continue;
-    const entry = presence[name];
-    const identity = identityByName.get(name);
-    const participant = participantByName.get(name);
-    // Preserve the existing presence → identity authority for current runtime
-    // metadata. participants fills the fresh-connection gap but never lets a
-    // connection snapshot overwrite an already known presence/identity row.
-    const kind = entry?.kind ?? identity?.kind ?? participant?.kind ?? "agent";
-    const account = entry?.account ?? identity?.account ?? participant?.owner;
-    const participantDisplay = participant?.handle
-      ?? (participant?.kind === "human" ? participant.display_name ?? participant.owner : undefined);
-    const display = identity?.display ?? participantDisplay ?? name;
-    const accountLabel = account && account !== ""
-      ? account
-      : kind === "human"
-        ? display
-        : t("Channel.roles.unowned");
-    const owner = account && account !== display ? account : null;
-    members.push({
-      name,
-      display,
-      accountLabel,
-      owner,
-      kind,
-    });
-  }
-  return members;
-}
-
-export function teamRoleBuckets(
-  assignedRoles: ChannelRoleInfo[],
-  presence: Record<string, PresenceEntry>,
-  identities: ChannelIdentity[],
-  participants: Sender[],
-  t: TFunc,
-): {
-  selfReported: ChannelRoleInfo[];
-  unassigned: UnassignedMember[];
-} {
-  const selfReported = selfReportedRoles(assignedRoles, presence, identities);
-  return {
-    selfReported,
-    unassigned: unassignedMembers(assignedRoles, selfReported, presence, identities, participants, t),
-  };
-}
-
 function CharterBanner({
   charter,
   open,
@@ -644,578 +514,6 @@ function CharterBanner({
         </div>
       )}
     </section>
-  );
-}
-
-export interface DivisionBoardProps {
-  canModerate: boolean;
-  slug: string;
-  roles: ChannelRoleInfo[];
-  roleDrafts: Record<string, RoleDraft>;
-  roleError: string | null;
-  roleSaving: string | null;
-  roleName: string;
-  roleDraft: RoleDraft;
-  identities: ChannelIdentity[];
-  presence: Record<string, PresenceEntry>;
-  participants: Sender[];
-  onlineNames?: ReadonlySet<string>;
-  onRoleDraft: (name: string, draft: RoleDraft) => void;
-  onNewRoleName: (name: string) => void;
-  onNewRoleDraft: (draft: RoleDraft) => void;
-  onSaveRole: (name: string, draft: RoleDraft) => Promise<boolean>;
-  onSetReportsTo?: (name: string, reportsTo: string | null) => void;
-  onDeleteRole: (name: string) => void;
-  forceOpen?: boolean;
-  // issue #150：一键把当前已声明分工同步进公告——DivisionBoard 只负责拼内容，
-  // 落盘（网络请求 + rev 刷新）交给上层（Channel.tsx 已有 charter 状态机）。
-  charterText: string | null;
-  onSyncToCharter: (text: string) => void;
-  syncingCharter: boolean;
-  // issue #171：分工面板到「查看/编辑每个 agent 自己的规则」（AgentTokens，已在
-  // commit 7f7e8e1 落地）的入口——门禁复用 Channel.tsx 里 AgentTokens 本身的门禁
-  // （canMintAgent && accountKey !== null），不在这里重新定义一套。
-  canManageAgentRules: boolean;
-  manageableAgentAccount?: string | null;
-  onOpenAgentRules: () => void;
-  onOpenAgentRulesFor?: (name: string) => void;
-  // issue #272（审计重开）：点分工面板里的某个成员，打开它的单 Agent 详情弹窗
-  // （工作状态/历史工作内容/在线状态）。可选——未接的调用方（如既有测试）行内保持只读展示。
-  onOpenAgentDetail?: (name: string) => void;
-}
-
-export function DivisionBoard({
-  canModerate,
-  slug,
-  roles,
-  roleDrafts,
-  roleError,
-  roleSaving,
-  roleName,
-  roleDraft,
-  identities,
-  presence,
-  participants,
-  onlineNames,
-  onRoleDraft,
-  onNewRoleName,
-  onNewRoleDraft,
-  onSaveRole,
-  onSetReportsTo,
-  onDeleteRole,
-  forceOpen = false,
-  charterText,
-  onSyncToCharter,
-  syncingCharter,
-  canManageAgentRules,
-  manageableAgentAccount,
-  onOpenAgentRules,
-  onOpenAgentRulesFor,
-  onOpenAgentDetail,
-}: DivisionBoardProps) {
-  const t = useT();
-  const [selfHintOpen, setSelfHintOpen] = useState(false);
-  const [selfHintCopied, setSelfHintCopied] = useState(false);
-  const [editingRoleName, setEditingRoleName] = useState<string | null>(null);
-  const editingRoleNameRef = useRef<string | null>(null);
-  const editableRolesRef = useRef<Map<string, ChannelRoleInfo>>(new Map());
-  const onRoleDraftRef = useRef(onRoleDraft);
-  const editButtonRefs = useRef(new Map<string, HTMLButtonElement>());
-  onRoleDraftRef.current = onRoleDraft;
-  const [unassignedOpen, setUnassignedOpen] = useState(false);
-  // #504 还原度：组织架构树默认折叠（设计里是个按钮，不是常驻大卡），点开才展开，别顶乱分工列表。
-  // 成员页首先回答「频道里有哪些人、每个人有哪些 agent」；汇报关系只是次级信息。
-  // 默认展开目录，避免用户进入 Members 后还要猜一次“组织架构”按钮才看得到 roster。
-  const [orgOpen, setOrgOpen] = useState(true);
-  useEffect(() => {
-    if (!selfHintCopied) return;
-    const timer = window.setTimeout(() => setSelfHintCopied(false), 1400);
-    return () => window.clearTimeout(timer);
-  }, [selfHintCopied]);
-  const selfReportCmd = `party status --channel ${slug} working --role worker --note ${JSON.stringify(
-    t("Channel.roles.selfReport.exampleNote"),
-  )}`;
-  const copySelfReportCmd = () => {
-    if (navigator.clipboard !== undefined) {
-      void navigator.clipboard
-        .writeText(selfReportCmd)
-        .then(() => setSelfHintCopied(true))
-        .catch(() => undefined);
-    }
-  };
-  const identityByName = new Map(identities.map((identity) => [identity.name, identity]));
-  const roleBuckets = teamRoleBuckets(roles, presence, identities, participants, t);
-  const selfRoles = roleBuckets.selfReported;
-  editableRolesRef.current = new Map(
-    [...roles, ...selfRoles].map((role) => [role.name, role]),
-  );
-  const resetRoleDraft = (name: string) => {
-    const source = editableRolesRef.current.get(name);
-    if (source !== undefined) onRoleDraftRef.current(name, roleDraftFrom(source));
-  };
-  const restoreEditButtonFocus = (name: string) => {
-    globalThis.setTimeout(() => editButtonRefs.current.get(name)?.focus(), 0);
-  };
-  const cancelRoleEdit = (name: string, restoreFocus = true) => {
-    if (roleSaving !== null) return;
-    resetRoleDraft(name);
-    editingRoleNameRef.current = null;
-    setEditingRoleName(null);
-    if (restoreFocus) restoreEditButtonFocus(name);
-  };
-  const beginRoleEdit = (name: string) => {
-    if (roleSaving !== null) return;
-    const previous = editingRoleNameRef.current;
-    if (previous !== null && previous !== name) resetRoleDraft(previous);
-    editingRoleNameRef.current = name;
-    setEditingRoleName(name);
-  };
-  const finishRoleEdit = (name: string) => {
-    editingRoleNameRef.current = null;
-    setEditingRoleName(null);
-    restoreEditButtonFocus(name);
-  };
-  useEffect(() => () => {
-    const name = editingRoleNameRef.current;
-    if (name !== null) resetRoleDraft(name);
-  }, []);
-  // issue #169：unassigned 也并入同一份 roleViews，保证 roster 完整。header 的
-  // 「N 个分工」只数正式 channel_roles；presence self report 仍显示，但标为待确认。
-  const unassigned = roleBuckets.unassigned;
-  // 只有 channel_roles 是正式分工；presence 自报只是一条待确认 claim。
-  const declaredCount = roles.length;
-  const roleViews = [
-    ...roles.map((role) => ({ ...roleViewFor(role, identityByName.get(role.name), t), source: "assigned" as const, name: role.name })),
-    ...selfRoles.map((role) => ({ ...roleViewFor(role, identityByName.get(role.name), t), source: "self" as const, name: role.name })),
-    ...unassigned.map((member) => ({ ...member, role: null, source: "unassigned" as const })),
-  ]
-    .sort(
-      (a, b) =>
-        a.accountLabel.localeCompare(b.accountLabel) ||
-        (a.role?.role ?? "\uffff").localeCompare(b.role?.role ?? "\uffff") ||
-        a.display.localeCompare(b.display),
-    );
-  // issue #168 + #370：只有 channel_roles assignment 是正式组织权威。
-  // presence.lineage 仍可在运行时详情里解释派生关系，但不能提升成角色、负责人或汇报线。
-  const knownAssignedNames = new Set(
-    roleViews.filter((view) => view.source === "assigned").map((view) => view.name),
-  );
-  const roleViewsWithReports = roleViews.map((view) => ({
-    ...view,
-    reportsTo: view.source === "assigned" ? view.role?.reports_to ?? null : null,
-  }));
-  // self/unassigned 仍作为 roster 节点显示，但不携带正式 role/reportsTo/lead 语义。
-  const orgMembers: OrgMemberInput[] = roleViewsWithReports.map((view) => ({
-    name: view.name,
-    display: view.display,
-    role: view.source === "assigned" ? view.role?.role ?? null : null,
-    reportsTo: view.reportsTo,
-    kind: view.kind,
-    accountLabel: view.accountLabel,
-    source: view.source,
-  }));
-  const orgTree = buildOrgTree(orgMembers);
-  const assignedRoleViews = roleViewsWithReports.filter((view) => view.role !== null);
-  const groups: Array<{ accountLabel: string; roles: typeof roleViewsWithReports }> = [];
-  for (const view of assignedRoleViews) {
-    const current = groups.at(-1);
-    if (current !== undefined && current.accountLabel === view.accountLabel) current.roles.push(view);
-    else groups.push({ accountLabel: view.accountLabel, roles: [view] });
-  }
-  const roleCounts = COLLAB_ROLES
-    .map((role) => ({
-      role,
-      count: roleViews.filter((item) => item.source === "assigned" && item.role?.role === role).length,
-    }))
-    .filter((item) => item.count > 0);
-
-  // issue #150：拿当前正式分工（不含 self report 与未分工占位）拼成 markdown
-  // 小节，合并进现有公告文本。这里只计算候选文本；写入必须由 moderator 明确点击
-  // 「同步到公告」触发，挂载 Team、切换页签或刷新角色数据都不能产生隐式写操作。
-  const declaredRoleIdentities = roleViews
-    .filter(
-      (view): view is typeof view & { role: NonNullable<typeof view.role> } =>
-        view.source === "assigned" && view.role !== null,
-    )
-    .map((view) => ({
-      ...view,
-      kind: view.role.kind ?? identityByName.get(view.name)?.kind ?? presence[view.name]?.kind,
-      roleName: view.role.role,
-    }));
-  const declared: DivisionCharterRole[] = declaredAgentRoles(
-    declaredRoleIdentities.map((view) => ({
-      name: view.name,
-      display: view.display,
-      kind: view.kind,
-      role: view.roleName,
-      source: view,
-    })),
-  )
-    .map((view) => ({
-      display: view.source.display,
-      accountLabel: view.source.accountLabel,
-      role: view.source.role.role,
-      responsibility: view.source.role.responsibility,
-    }));
-  const currentCharterText = charterText ?? "";
-  const nextCharterText = mergeDivisionIntoCharter(
-    currentCharterText,
-    formatDivisionSection(declared, {
-      heading: t("Channel.roles.syncHeading"),
-      empty: t("Channel.roles.syncEmpty"),
-    }),
-  );
-  const syncDivisionToCharter = () => {
-    if (charterText === null) return;
-    onSyncToCharter(nextCharterText);
-  };
-
-  return (
-    <details className="role-board" aria-label={t("Channel.roles.label")} open={forceOpen ? true : undefined}>
-      <summary className="role-board-head">
-        <div>
-          <h2>{t("Channel.roles.label")}</h2>
-          <p className="t-mono">{t("Channel.roles.help")}</p>
-        </div>
-        <div className="role-board-summary">
-          <span className="t-mono role-board-count">{t("Channel.roles.count", { count: String(declaredCount) })}</span>
-          {roleCounts.map((item) => (
-            <span key={item.role} className="t-mono role-board-role-count">
-              {roleCountLabel(item.role, item.count, t)}
-            </span>
-          ))}
-        </div>
-      </summary>
-      <div className="role-board-body">
-        <div className="role-board-actions">
-          {canModerate && (
-            <button
-              type="button"
-              className="d-btn role-sync-charter-btn"
-              disabled={syncingCharter || charterText === null}
-              onClick={syncDivisionToCharter}
-            >
-              {syncingCharter ? t("Channel.roles.syncingCharter") : t("Channel.roles.syncToCharter")}
-            </button>
-          )}
-          {canManageAgentRules && (
-            <button type="button" className="d-btn role-open-rules-btn" onClick={onOpenAgentRules}>
-              {t("Channel.roles.openAgentRules")}
-            </button>
-          )}
-          <button
-            type="button"
-            className="d-btn role-org-toggle"
-            aria-expanded={orgOpen}
-            aria-controls="division-org-tree"
-            onClick={() => setOrgOpen((v) => !v)}
-          >
-            {t("Channel.roles.orgToggle")} <span aria-hidden="true">{orgOpen ? "▾" : "▸"}</span>
-          </button>
-        </div>
-        <div className="role-selfhint">
-          <button
-            type="button"
-            className="role-selfhint-toggle t-mono"
-            aria-expanded={selfHintOpen}
-            onClick={() => setSelfHintOpen((v) => !v)}
-          >
-            <span>{t("Channel.roles.selfReport.summary")}</span>
-            <span className="role-selfhint-arrow" aria-hidden="true">{selfHintOpen ? "▾" : "▸"}</span>
-          </button>
-          {selfHintOpen && (
-            <div className="role-selfhint-body">
-              <p>{t("Channel.roles.selfReport.intro")}</p>
-              <div className="role-selfhint-cmd">
-                <code className="t-mono">{selfReportCmd}</code>
-                <button type="button" className="d-btn role-selfhint-copy" onClick={copySelfReportCmd}>
-                  {selfHintCopied ? t("Channel.roles.selfReport.copied") : t("Channel.roles.selfReport.copy")}
-                </button>
-              </div>
-              <p className="t-mono role-selfhint-meta">{t("Channel.roles.selfReport.roles")}</p>
-              <p className="role-selfhint-caveat">{t("Channel.roles.selfReport.hostCaveat")}</p>
-            </div>
-          )}
-        </div>
-        {orgOpen && (
-          <OrgTreePreview
-            id="division-org-tree"
-            tree={orgTree}
-            t={t}
-            interactive={{
-              canModerate,
-              allNames: orgMembers
-                .filter((member) => member.source === "assigned")
-                .map((member) => member.name),
-              busyName: roleSaving,
-              onSetReportsTo: onSetReportsTo ?? (() => {}),
-              onOpenMember: onOpenAgentDetail,
-              onManageAgent: canManageAgentRules ? onOpenAgentRulesFor : undefined,
-              manageableAccount: manageableAgentAccount,
-            }}
-          />
-        )}
-        {groups.length > 0 ? (
-          <div className="role-account-list">
-            {groups.map((group) => (
-              <section key={group.accountLabel} className="role-account-group">
-                <header className="role-account-head">
-                  <span className="role-account-label">{group.accountLabel}</span>
-                  <span className="t-mono role-account-count">
-                    {t("Channel.roles.accountCount", { count: String(group.roles.length) })}
-                  </span>
-                </header>
-                <div className="role-list">
-                  {group.roles.map(({ role, display, owner, accountLabel, kind, source, name, reportsTo }) => {
-                    // issue #169：role 为 null 代表「已连接/曾出现过，但从没声明过角色」的
-                    // 未分工成员——只读展示占位文案，不接可编辑的 role-select/input（那需要一个
-                    // 真实的 ChannelRoleInfo；moderator 要给他分工，走下面的「添加」新建行）。
-                    const draftForRole = role !== null ? roleDrafts[role.name] ?? roleDraftFrom(role) : null;
-                    const savingThisRole =
-                      role !== null
-                      && (roleSaving === role.name || (source === "self" && roleSaving === "__new__"));
-                    // issue #168：汇报对象是否在本频道 roster 里可见——不可见就提示，帮助
-                    // 发现/避免跨出本频道边界的汇报关系。
-                    const reportsToVisible = reportsTo !== null && knownAssignedNames.has(reportsTo);
-                    const online = onlineNames?.has(name) ?? presence[name]?.live === true;
-                    const title = [
-                      name !== display ? name : null,
-                      t("Composer.owner", { account: accountLabel }),
-                      t(`Composer.kind.${kind}`),
-                      role !== null ? t("Composer.role", { role: role.role }) : null,
-                      role?.responsibility ? t("Composer.responsibility", { responsibility: role.responsibility }) : null,
-                    ].filter((part): part is string => part !== null).join("\n");
-                    return (
-                      <div
-                        key={name}
-                        className={`role-row role-row--card role-row--${role?.role ?? "unassigned"}`}
-                        aria-busy={editingRoleName === role?.name && roleSaving !== null}
-                        onKeyDown={(event) => {
-                          if (event.key !== "Escape" || role === null || editingRoleName !== role.name) return;
-                          event.preventDefault();
-                          event.stopPropagation();
-                          cancelRoleEdit(role.name);
-                        }}
-                      >
-                        <div className="role-person" title={title}>
-                          <span
-                            className={`role-live-dot${online ? " is-online" : ""}`}
-                            role="img"
-                            aria-label={t(online ? "Channel.team.member.online" : "Channel.team.member.offline")}
-                          />
-                          {onOpenAgentDetail !== undefined ? (
-                            <button
-                              type="button"
-                              className="role-person-name role-person-name--btn t-mono"
-                              data-team-member={name}
-                              onClick={() => onOpenAgentDetail(name)}
-                            >
-                              {display}
-                            </button>
-                          ) : (
-                            <span className="role-person-name t-mono">{display}</span>
-                          )}
-                          <span className={`role-kind role-kind--${kind}`}>{t(`Composer.kind.${kind}`)}</span>
-                          {source === "assigned" && role !== null && role.role === "host" && (
-                            <span className="role-lead-tag t-mono">{t("Channel.roles.channelLead")}</span>
-                          )}
-                          {source === "self" && <span className="role-source t-mono">{t("Channel.roles.selfReported")}</span>}
-                          {source === "unassigned" && (
-                            <span className="role-source role-source--unassigned t-mono">{t("Channel.roles.unassigned")}</span>
-                          )}
-                          {owner !== null && <span className="role-owner t-mono">{owner}</span>}
-                          {reportsTo !== null && (
-                            <span
-                              className={"role-report" + (reportsToVisible ? "" : " role-report--external") + " t-mono"}
-                            >
-                              {reportsToVisible
-                                ? t("Channel.roles.reportsTo", { parent: reportsTo })
-                                : t("Channel.roles.reportsToExternal", { parent: reportsTo })}
-                            </span>
-                          )}
-                        </div>
-                        {role === null || draftForRole === null ? (
-                          <span className="role-text role-text--unassigned">{t("Channel.roles.noRoleYet")}</span>
-                        ) : canModerate && editingRoleName === role.name ? (
-                          <>
-                            <select
-                              className="role-select t-mono"
-                              aria-label={t("Channel.roles.roleLabel")}
-                              value={draftForRole.role}
-                              disabled={roleSaving !== null}
-                              autoFocus
-                              onChange={(e) => onRoleDraft(role.name, { ...draftForRole, role: e.target.value as CollaborationRole })}
-                            >
-                              {COLLAB_ROLES.map((item) => (
-                                <option key={item} value={item}>{item}</option>
-                              ))}
-                            </select>
-                            <input
-                              className="role-input"
-                              aria-label={t("Channel.roles.responsibilityLabel")}
-                              value={draftForRole.responsibility}
-                              disabled={roleSaving !== null}
-                              onChange={(e) => onRoleDraft(role.name, { ...draftForRole, responsibility: e.target.value })}
-                              autoComplete="off"
-                              placeholder={t("Channel.roles.responsibilityPlaceholder")}
-                            />
-                            <button
-                              className="d-btn"
-                              type="button"
-                              disabled={roleSaving !== null}
-                              onClick={() => {
-                                void onSaveRole(role.name, draftForRole)
-                                  .then((saved) => {
-                                    if (saved) finishRoleEdit(role.name);
-                                  })
-                                  .catch(() => undefined);
-                              }}
-                            >
-                              {savingThisRole
-                                ? t("Channel.roles.saving")
-                                : source === "self"
-                                  ? t("Channel.roles.saveAndConfirm")
-                                  : t("Channel.roles.save")}
-                            </button>
-                            <button
-                              className="d-btn"
-                              type="button"
-                              disabled={roleSaving !== null}
-                              onClick={() => cancelRoleEdit(role.name)}
-                            >
-                              {t("Channel.roles.cancel")}
-                            </button>
-                            {source === "assigned" && (
-                              <button className="d-btn" type="button" disabled={roleSaving !== null} onClick={() => onDeleteRole(role.name)}>
-                                {t("Channel.roles.clear")}
-                              </button>
-                            )}
-                          </>
-                        ) : (
-                          <>
-                            <span className="role-badge t-mono">{role.role}</span>
-                            <span className="role-text">{role.responsibility ?? t("Channel.roles.noResponsibility")}</span>
-                            {canModerate && (
-                              <button
-                                className="d-btn role-edit-btn"
-                                type="button"
-                                data-role-edit={role.name}
-                                ref={(node) => {
-                                  if (node === null) editButtonRefs.current.delete(role.name);
-                                  else editButtonRefs.current.set(role.name, node);
-                                }}
-                                disabled={roleSaving !== null}
-                                onClick={() => beginRoleEdit(role.name)}
-                              >
-                                {source === "self" ? t("Channel.roles.confirm") : t("Channel.roles.edit")}
-                              </button>
-                            )}
-                          </>
-                        )}
-                      </div>
-                    );
-                  })}
-                </div>
-              </section>
-            ))}
-          </div>
-        ) : (
-          <p className="charter-empty">{t("Channel.roles.empty")}</p>
-        )}
-        {unassigned.length > 0 && (
-          <section className="role-unassigned-fold">
-            <button
-              type="button"
-              className="role-unassigned-toggle t-mono"
-              aria-expanded={unassignedOpen}
-              onClick={() => setUnassignedOpen((open) => !open)}
-            >
-              <span aria-hidden="true">{unassignedOpen ? "▾" : "▸"}</span>
-              <span>
-                {t(canModerate ? "Channel.roles.unassignedFold" : "Channel.roles.unassignedFoldReadonly", {
-                  count: String(unassigned.length),
-                })}
-              </span>
-            </button>
-            {unassignedOpen && (
-              <div className="role-unassigned-chips">
-                {unassigned.map((member) => canModerate ? (
-                  <button
-                    key={member.name}
-                    type="button"
-                    className="role-unassigned-chip role-person-name role-person-name--btn t-mono"
-                    title={member.owner ?? member.accountLabel}
-                    disabled={roleSaving !== null}
-                    onClick={() => onNewRoleName(member.name)}
-                  >
-                    {member.display}
-                  </button>
-                ) : (
-                  <span
-                    key={member.name}
-                    className="role-unassigned-chip role-person-name t-mono"
-                    title={member.owner ?? member.accountLabel}
-                  >
-                    {member.display}
-                  </span>
-                ))}
-              </div>
-            )}
-          </section>
-        )}
-        {canModerate && (
-          <div className="role-row role-row--new">
-            <input
-              className="role-name-input t-mono"
-              aria-label={t("Channel.roles.nameLabel")}
-              value={roleName}
-              disabled={roleSaving !== null}
-              onChange={(e) => onNewRoleName(e.target.value)}
-              list="channel-role-targets"
-              autoComplete="off"
-              spellCheck={false}
-              placeholder={t("Channel.roles.namePlaceholder")}
-            />
-            <select
-              className="role-select t-mono"
-              aria-label={t("Channel.roles.roleLabel")}
-              value={roleDraft.role}
-              disabled={roleSaving !== null}
-              onChange={(e) => onNewRoleDraft({ ...roleDraft, role: e.target.value as CollaborationRole })}
-            >
-              {COLLAB_ROLES.map((item) => (
-                <option key={item} value={item}>{item}</option>
-              ))}
-            </select>
-            <input
-              className="role-input"
-              aria-label={t("Channel.roles.responsibilityLabel")}
-              value={roleDraft.responsibility}
-              disabled={roleSaving !== null}
-              onChange={(e) => onNewRoleDraft({ ...roleDraft, responsibility: e.target.value })}
-              autoComplete="off"
-              placeholder={t("Channel.roles.responsibilityPlaceholder")}
-            />
-            <button
-              className="d-btn d-btn--primary"
-              type="button"
-              disabled={roleSaving !== null}
-              onClick={() => {
-                void onSaveRole(roleName, roleDraft).catch(() => undefined);
-              }}
-            >
-              {roleSaving === "__new__" ? t("Channel.roles.saving") : t("Channel.roles.add")}
-            </button>
-            <datalist id="channel-role-targets">
-              {identities.map((identity) => (
-                <option key={identity.name} value={identity.name}>{identity.display}</option>
-              ))}
-            </datalist>
-          </div>
-        )}
-        {roleError !== null && <p className="banner banner--red" role="alert">{roleError}</p>}
-      </div>
-    </details>
   );
 }
 
@@ -1729,472 +1027,6 @@ export function CompletionPanel({
           );
         })}
       </ol>}
-    </section>
-  );
-}
-
-export function TeamPanel({ teams }: { teams: TeamSummary[] }) {
-  const t = useT();
-  const [expandedMember, setExpandedMember] = useState<string | null>(null);
-  if (teams.length === 0) return null;
-
-  return (
-    <section className="team-panel" aria-label={t("Channel.team.aria")}>
-      <div className="team-panel-head">
-        <h2 className="team-title">{t("Channel.heading.teams")}</h2>
-        <span className="t-mono team-count">{teams.length}</span>
-      </div>
-      <ol className="team-list">
-        {teams.map((team) => {
-          const front = team.frontAgent;
-          const workerMembers = front === null ? team.members : team.members.filter((member) => member.name !== front.name);
-          const meta = [
-            t("Channel.team.meta.root", { name: team.rootAgent }),
-            team.parentAgents.length === 1
-              ? t("Channel.team.meta.parent", { name: team.parentAgents[0]! })
-              : t("Channel.team.meta.parents", { count: team.parentAgents.length }),
-            t("Channel.team.meta.depth", { depth: team.maxDepth }),
-            team.expiresAt !== null ? t("Channel.team.meta.expires", { time: fmtTime(team.expiresAt) }) : null,
-            team.lastSeen !== null ? t("Channel.team.meta.seen", { time: fmtTime(team.lastSeen) }) : null,
-          ].filter((part): part is string => part !== null);
-          return (
-            <li key={team.key} className="team-item">
-              <div className="team-item-head">
-                <span className="team-name">{team.teamId}</span>
-                <span
-                  className={"t-mono team-front" + (front?.active ? " is-active" : "")}
-                  title={front === null
-                    ? t("Channel.team.front", { name: team.rootAgent })
-                    : t("Channel.team.frontTitle", {
-                        name: front.name,
-                        state: presenceStateLabel(front.state, t),
-                        residency: residencyLabel(front.residency, t),
-                      })}
-                >
-                  <span className={`d-dot d-dot--${front?.active ? front.state : "offline"}`} />
-                  {t("Channel.team.front", { name: front?.name ?? team.rootAgent })}
-                </span>
-                <span className="t-mono team-active">
-                  {t("Channel.team.active", { active: team.activeCount, total: team.memberCount })}
-                </span>
-                <span className={`t-mono team-residency team-residency--${team.residency}`}>
-                  {residencyLabel(team.residency, t)}
-                </span>
-              </div>
-              <div className="t-mono team-meta">{meta.join(" · ")}</div>
-              <div className="team-members">
-                {workerMembers.length === 0 && <span className="t-mono team-member team-member--empty">{t("Channel.team.noWorkers")}</span>}
-                {workerMembers.map((member) => {
-                  const detail = [
-                    t("Channel.team.memberTitle", {
-                      name: member.name,
-                      parent: member.parentAgent,
-                      state: presenceStateLabel(member.state, t),
-                      residency: residencyLabel(member.residency, t),
-                    }),
-                    member.expiresAt !== null ? t("Channel.team.meta.expires", { time: fmtTime(member.expiresAt) }) : null,
-                    member.lastSeen !== null ? t("Channel.team.meta.seen", { time: fmtTime(member.lastSeen) }) : null,
-                  ].filter((part): part is string => part !== null).join(" · ");
-                  const expanded = expandedMember === member.name;
-                  return (
-                    <Fragment key={member.name}>
-                      <button
-                        type="button"
-                        className={"t-mono team-member" + (member.active ? " is-active" : "")}
-                        title={detail}
-                        aria-expanded={expanded}
-                        onClick={() => setExpandedMember((current) => current === member.name ? null : member.name)}
-                      >
-                        <span className={`d-dot d-dot--${member.active ? member.state : "offline"}`} />
-                        <span>{member.name}</span>
-                      </button>
-                      {expanded && <span className="t-mono team-member-detail">{detail}</span>}
-                    </Fragment>
-                  );
-                })}
-              </div>
-            </li>
-          );
-        })}
-      </ol>
-    </section>
-  );
-}
-
-// #187：agent 维度看板——每个 agent 在忙/空闲/阻塞/离线，手里在做/排队/待审/受阻多少任务。
-// 纯只读聚合：合并 presence、task 台账与可靠投递状态；delivery/current_task 反指消息正文，直接说明在忙什么。
-type AgentBoardStatus = "busy" | "blocked" | "idle" | "offline";
-const AGENT_STATUS_ORDER: Record<AgentBoardStatus, number> = { busy: 0, blocked: 1, idle: 2, offline: 3 };
-const ACTIVE_DELIVERY_STATES = new Set<PublicDirectedDelivery["state"]>(["queued", "claimed", "running", "waiting_owner"]);
-// 一个 agent 积压几十条排队投递时，卡片不能被撑成整屏——只展示最近几条，其余折叠成一行计数。
-const AGENT_BOARD_MAX_WORK_ROWS = 5;
-
-export function activeDeliveryTargets(deliveries: PublicDirectedDelivery[]): string[] {
-  return [...new Set(deliveries.filter((delivery) => ACTIVE_DELIVERY_STATES.has(delivery.state)).map((delivery) => delivery.target_name))];
-}
-
-function agentWorkSummary(body: string): string {
-  const compact = body.replace(/\s+/g, " ").trim();
-  return compact.length > 160 ? `${compact.slice(0, 157)}…` : compact;
-}
-
-// 已撤回消息的本地行是 [retracted] 占位甚至历史残留，正文和 delivery.preview 都不可再展示。
-function agentWorkSummaryFor(message: MsgFrame | undefined, preview?: string | null): string {
-  if (message?.retracted === true) return "";
-  return agentWorkSummary(message?.body ?? "") || (preview ?? "");
-}
-
-export function agentPresenceSummary(
-  presence: PresenceEntry[],
-  participants: Sender[],
-  taskAgentNames: Iterable<string> = [],
-  memberNames?: ReadonlySet<string>,
-) {
-  const included = (name: string) => memberNames === undefined || memberNames.has(name);
-  const participantKinds = new Map(participants.map((participant) => [participant.name, participant.kind]));
-  const agentNames = new Set<string>();
-  for (const entry of presence) {
-    if (included(entry.name) && (participantKinds.get(entry.name) ?? entry.kind) !== "human") agentNames.add(entry.name);
-  }
-  for (const participant of participants) {
-    if (!included(participant.name)) continue;
-    if (participant.kind === "agent") agentNames.add(participant.name);
-    else agentNames.delete(participant.name);
-  }
-  for (const name of taskAgentNames) {
-    if (included(name)) agentNames.add(name);
-  }
-
-  // participants is the Durable Object's authoritative list of currently open
-  // connections. A freshly connected agent does not necessarily have a
-  // persisted presence row yet, so presence.live alone under-counts the board.
-  const onlineNames = new Set(
-    participants
-      .filter((participant) => participant.kind === "agent" && included(participant.name))
-      .map((participant) => participant.name),
-  );
-  for (const entry of presence) {
-    if (
-      included(entry.name)
-      && entry.live === true
-      && (participantKinds.get(entry.name) ?? entry.kind) !== "human"
-    ) {
-      onlineNames.add(entry.name);
-    }
-  }
-  return { agentNames, onlineNames, online: onlineNames.size, offline: agentNames.size - onlineNames.size };
-}
-
-export function teamMemberOnlineNames(
-  presence: PresenceEntry[],
-  participants: Sender[],
-): Set<string> {
-  const onlineNames = new Set(participants.map((participant) => participant.name));
-  for (const entry of presence) {
-    if (entry.live === true) onlineNames.add(entry.name);
-  }
-  return onlineNames;
-}
-
-export function teamMemberPresenceSummary(
-  memberNames: ReadonlySet<string>,
-  onlineNames: ReadonlySet<string>,
-): { online: number; offline: number } {
-  let online = 0;
-  for (const name of memberNames) {
-    if (onlineNames.has(name)) online += 1;
-  }
-  return { online, offline: memberNames.size - online };
-}
-
-export function agentBoardTaskAssignee(task: TaskRecord): string | null {
-  const name = task.assignee?.name;
-  if (!name || task.assignee?.kind !== "agent") return null;
-  return task.state === "in_progress" || task.state === "assigned" || task.state === "needs_review" || task.state === "blocked"
-    ? name
-    : null;
-}
-
-export function AgentBoardPanel({
-  presence,
-  participants = [],
-  tasks,
-  deliveries = [],
-  messages = [],
-  onOpenAgentDetail,
-  onOpenTask,
-  onOpenMessage,
-  memberNames,
-}: {
-  presence: PresenceEntry[];
-  participants?: Sender[];
-  tasks: TaskRecord[];
-  deliveries?: PublicDirectedDelivery[];
-  messages?: MsgFrame[];
-  onOpenAgentDetail?: (name: string) => void;
-  onOpenTask?: (id: number) => void;
-  onOpenMessage?: (seq: number) => void | Promise<void>;
-  /** Current channel roster; task/delivery/history cannot recreate a removed member. */
-  memberNames?: ReadonlySet<string>;
-}) {
-  const t = useT();
-  const messagesBySeq = new Map(messages.map((message) => [message.seq, message]));
-  const activeDeliveriesByName = new Map<string, PublicDirectedDelivery[]>();
-  for (const delivery of deliveries) {
-    if (memberNames !== undefined && !memberNames.has(delivery.target_name)) continue;
-    if (!ACTIVE_DELIVERY_STATES.has(delivery.state)) continue;
-    const assigned = activeDeliveriesByName.get(delivery.target_name) ?? [];
-    assigned.push(delivery);
-    activeDeliveriesByName.set(delivery.target_name, assigned);
-  }
-  for (const assigned of activeDeliveriesByName.values()) {
-    assigned.sort((a, b) => b.updated_at - a.updated_at || b.message_seq - a.message_seq || a.id.localeCompare(b.id));
-  }
-  const tasksByName = new Map<string, TaskRecord[]>();
-  for (const task of tasks) {
-    const name = agentBoardTaskAssignee(task);
-    if (name === null) continue;
-    if (memberNames !== undefined && !memberNames.has(name)) continue;
-    const assigned = tasksByName.get(name) ?? [];
-    assigned.push(task);
-    tasksByName.set(name, assigned);
-  }
-  const taskStateOrder: Partial<Record<TaskRecord["state"], number>> = {
-    in_progress: 0,
-    blocked: 1,
-    needs_review: 2,
-    assigned: 3,
-  };
-  for (const assigned of tasksByName.values()) {
-    assigned.sort((a, b) =>
-      (taskStateOrder[a.state] ?? 99) - (taskStateOrder[b.state] ?? 99)
-      || b.priority - a.priority
-      || b.updated_at - a.updated_at
-      || a.id - b.id,
-    );
-  }
-  const presenceByName = new Map(presence.map((p) => [p.name, p]));
-  const presenceSummary = agentPresenceSummary(
-    presence,
-    participants,
-    [...tasksByName.keys(), ...activeDeliveriesByName.keys()],
-    memberNames,
-  );
-  const names = new Set(presenceSummary.agentNames);
-  const statusOf = (name: string, p: PresenceEntry | undefined, activeDeliveries: PublicDirectedDelivery[]): AgentBoardStatus => {
-    if (!presenceSummary.onlineNames.has(name)) return "offline";
-    if (p?.state === "blocked" || activeDeliveries.some((delivery) => delivery.state === "waiting_owner")) return "blocked";
-    if (
-      p?.busy === true
-      || p?.state === "working"
-      || activeDeliveries.some((delivery) => delivery.state === "queued" || delivery.state === "claimed" || delivery.state === "running")
-    ) return "busy";
-    return "idle";
-  };
-  const rows = [...names]
-    .map((name) => {
-      const p = presenceByName.get(name);
-      const assigned = tasksByName.get(name) ?? [];
-      const activeDeliveries = activeDeliveriesByName.get(name) ?? [];
-      const knownDeliverySeqs = new Set(activeDeliveries.map((delivery) => delivery.message_seq));
-      const activeWork = activeDeliveries.map((delivery) => ({
-        id: delivery.id,
-        seq: delivery.message_seq,
-        state: delivery.state,
-        // 优先本地已加载正文（含实时编辑），窗口外的老消息退回 worker 投影时带的 preview。
-        summary: agentWorkSummaryFor(messagesBySeq.get(delivery.message_seq), delivery.preview),
-      }));
-      if (typeof p?.current_task === "number" && !knownDeliverySeqs.has(p.current_task)) {
-        activeWork.unshift({
-          id: `presence-${name}-${p.current_task}`,
-          seq: p.current_task,
-          state: "running",
-          summary: agentWorkSummaryFor(messagesBySeq.get(p.current_task)),
-        });
-      }
-      // #187 第4项「排期」：surface presence 里的暂停/定时恢复（resume_at），看板本行直接可见。
-      return {
-        name,
-        status: statusOf(name, p, activeDeliveries),
-        online: presenceSummary.onlineNames.has(name),
-        note: p?.note ?? null,
-        paused: p?.paused === true,
-        resumeAt: p?.resume_at ?? null,
-        activeWork,
-        tasks: assigned,
-        inProgress: assigned.filter((task) => task.state === "in_progress").length,
-        queued: assigned.filter((task) => task.state === "assigned").length,
-        review: assigned.filter((task) => task.state === "needs_review").length,
-        blocked: assigned.filter((task) => task.state === "blocked").length,
-      };
-    })
-    .sort((a, b) =>
-      AGENT_STATUS_ORDER[a.status] - AGENT_STATUS_ORDER[b.status]
-      || (a.status === "offline" && b.status === "offline"
-        ? Number(Boolean(b.note?.trim())) - Number(Boolean(a.note?.trim()))
-        : 0)
-      || b.inProgress - a.inProgress
-      || a.name.localeCompare(b.name),
-    );
-  const laneStatuses: AgentBoardStatus[] = ["busy", "blocked", "idle", "offline"];
-  const counts = Object.fromEntries(laneStatuses.map((status) => [status, rows.filter((row) => row.status === status).length])) as Record<AgentBoardStatus, number>;
-
-  if (rows.length === 0) {
-    return (
-      <section className="agent-board-panel" aria-label={t("Channel.agentBoard.aria")}>
-        <p className="agent-board-empty">{t("Channel.agents.empty")}</p>
-      </section>
-    );
-  }
-  return (
-    <section className="agent-board-panel" aria-label={t("Channel.agentBoard.aria")}>
-      <header className="agent-board-overview">
-        <span className="t-mono agent-board-summary">
-          {t("Channel.agentBoard.summary", {
-            busy: String(counts.busy),
-            blocked: String(counts.blocked),
-            idle: String(counts.idle),
-            offline: String(counts.offline),
-          })}
-        </span>
-      </header>
-      {laneStatuses.map((status) => {
-        const laneRows = rows.filter((row) => row.status === status);
-        const statusLabel = t(`Channel.agents.status.${status}`);
-        const cards = (
-          <div className="agent-board-lane-cards">
-            {laneRows.length === 0 && <p className="agent-board-lane-empty">{statusLabel} · 0 — {t("Channel.tasks.columnEmpty")}</p>}
-            {laneRows.map((row) => (
-              <article key={row.name} className={`agent-board-row agent-board-row--${row.status}`} data-agent={row.name}>
-                <div className="agent-board-row-head">
-                  {onOpenAgentDetail !== undefined ? (
-                    <button
-                      type="button"
-                      className="agent-board-name agent-board-name--button"
-                      data-team-member={row.name}
-                      onClick={() => onOpenAgentDetail(row.name)}
-                    >
-                      <span className={`agent-board-live-dot${row.online ? " is-online" : ""}`} aria-hidden="true" />
-                      {row.name}
-                    </button>
-                  ) : (
-                    <span className="agent-board-name">
-                      <span className={`agent-board-live-dot${row.online ? " is-online" : ""}`} aria-hidden="true" />
-                      {row.name}
-                    </span>
-                  )}
-                  <span className={`t-mono agent-board-status agent-board-status--${row.status}`}>{statusLabel}</span>
-                </div>
-                {row.note !== null && row.note.trim() !== "" && <p className="agent-board-note">{row.note}</p>}
-                {row.paused && (
-                  <p className="t-mono agent-board-schedule">
-                    {row.resumeAt !== null
-                      ? t("Channel.agents.pausedUntil", { time: fmtTime(row.resumeAt) })
-                      : t("Channel.agents.pausedManual")}
-                  </p>
-                )}
-                <div className="agent-board-counts t-mono">
-                  <span title={t("Channel.agents.count.inProgress")}>▶ {row.inProgress}</span>
-                  <span title={t("Channel.agents.count.queued")}>⏳ {row.queued}</span>
-                  <span title={t("Channel.agents.count.review")}>👁 {row.review}</span>
-                  {row.blocked > 0 && <span className="agent-board-count-blocked" title={t("Channel.agents.count.blocked")}>⛔ {row.blocked}</span>}
-                </div>
-                {row.activeWork.length > 0 && (
-                  <ol className="agent-board-task-list agent-board-work-list">
-                    {row.activeWork.slice(0, AGENT_BOARD_MAX_WORK_ROWS).map((work) => (
-                      <li key={work.id} className={`agent-board-task agent-board-work agent-board-work--${work.state}`}>
-                        <span className="t-mono agent-board-task-id">#{work.seq}</span>
-                        {onOpenMessage !== undefined ? (
-                          <button
-                            type="button"
-                            className="agent-board-task-title agent-board-name--button"
-                            title={work.summary || undefined}
-                            aria-label={t("AgentDetailModal.openMessage", {
-                              seq: String(work.seq),
-                              summary: work.summary || t("Channel.agentBoard.workUnavailable"),
-                            })}
-                            onClick={() => { void onOpenMessage(work.seq); }}
-                          >
-                            {work.summary || t("Channel.agentBoard.workUnavailable")}
-                          </button>
-                        ) : (
-                          <span className="agent-board-task-title" title={work.summary || undefined}>
-                            {work.summary || t("Channel.agentBoard.workUnavailable")}
-                          </span>
-                        )}
-                        <span className="t-mono agent-board-task-state">{t(`WakeReceipt.delivery.state.${work.state}`)}</span>
-                      </li>
-                    ))}
-                    {row.activeWork.length > AGENT_BOARD_MAX_WORK_ROWS && (
-                      <li className="agent-board-task agent-board-work agent-board-work-more t-mono">
-                        {t("Channel.agentBoard.workMore", { count: String(row.activeWork.length - AGENT_BOARD_MAX_WORK_ROWS) })}
-                      </li>
-                    )}
-                  </ol>
-                )}
-                {row.tasks.length > 0 && (
-                  <ol className="agent-board-task-list">
-                    {row.tasks.map((task) => (
-                      <li key={task.id} className={`agent-board-task agent-board-task--${task.state}`}>
-                        <span className="t-mono agent-board-task-id">#{task.id}</span>
-                        {onOpenTask !== undefined ? (
-                          <button
-                            type="button"
-                            className="agent-board-task-title agent-board-name--button"
-                            title={task.title}
-                            aria-label={t("AgentDetailModal.openTask", {
-                              id: String(task.id),
-                              title: task.title,
-                            })}
-                            onClick={() => onOpenTask(task.id)}
-                          >
-                            {task.title}
-                          </button>
-                        ) : (
-                          <span className="agent-board-task-title" title={task.title}>{task.title}</span>
-                        )}
-                        <span className="t-mono agent-board-task-state">{t(`Channel.tasks.state.${task.state}`)}</span>
-                      </li>
-                    ))}
-                  </ol>
-                )}
-              </article>
-            ))}
-          </div>
-        );
-        if (status === "offline") {
-          const handoffCount = laneRows.filter((row) => row.note?.trim()).length;
-          return (
-            <details
-              key={status}
-              className="agent-board-lane agent-board-lane--offline"
-              data-status={status}
-            >
-              <summary className="agent-board-lane-head">
-                <h3 className="agent-board-lane-title agent-board-status--offline">{statusLabel}</h3>
-                <span className="t-mono agent-board-lane-count">{laneRows.length}</span>
-                {handoffCount > 0 && <span className="t-mono agent-board-handoff-count">{t("Channel.agentBoard.handoffs", { count: String(handoffCount) })}</span>}
-              </summary>
-              {cards}
-            </details>
-          );
-        }
-        return (
-          <section
-            key={status}
-            className={`agent-board-lane agent-board-lane--${status}`}
-            data-status={status}
-            data-empty={laneRows.length === 0}
-            aria-label={t("Channel.tasks.columnAria", { state: statusLabel })}
-          >
-            <header className="agent-board-lane-head">
-              <h3 className={`agent-board-lane-title agent-board-status--${status}`}>{statusLabel}</h3>
-              <span className="t-mono agent-board-lane-count">{laneRows.length}</span>
-            </header>
-            {cards}
-          </section>
-        );
-      })}
     </section>
   );
 }
